refactor: create pages folder to maintain pages organized
