+++ src/app/pages/home/home.component.html
@@ -40,3 +40,4 @@
   </div>
 
 </div>
+
