+++ src/app/pages/planning/planning.component.html
@@ -2,3 +2,4 @@
   <h2 class="text-2xl font-bold text-neutral-900 mb-2">Planejamento</h2>
   <p class="text-neutral-600">Em breve.</p>
 </section>
+
