+++ src/app/pages/register/register.component.html
@@ -156,3 +156,4 @@
     </div>
   </div>
 </div>
+
