+++ src/app/pages/settings/settings.component.ts
